Implemented NPC 63 (Toroko attacking)
Also refactored a lot of shit and made build faster
