+++ CaveStoryRemake.cbp
@@ -13,6 +13,93 @@
 				<Option compiler="null" />
 			</Target>
 		</Build>
+		<Unit filename="source/bullet.cpp" />
+		<Unit filename="source/bullet.h" />
+		<Unit filename="source/bulletCollision.cpp" />
+		<Unit filename="source/bulletCollision.h" />
+		<Unit filename="source/caret.cpp" />
+		<Unit filename="source/caret.h" />
+		<Unit filename="source/common.h" />
+		<Unit filename="source/fade.cpp" />
+		<Unit filename="source/fade.h" />
+		<Unit filename="source/filesystem.cpp" />
+		<Unit filename="source/filesystem.h" />
+		<Unit filename="source/fireball.cpp" />
+		<Unit filename="source/fireball.h" />
+		<Unit filename="source/fireballShoot.cpp" />
+		<Unit filename="source/flags.cpp" />
+		<Unit filename="source/flags.h" />
+		<Unit filename="source/game.cpp" />
+		<Unit filename="source/game.h" />
+		<Unit filename="source/hud.cpp" />
+		<Unit filename="source/hud.h" />
+		<Unit filename="source/input.cpp" />
+		<Unit filename="source/input.h" />
+		<Unit filename="source/level.cpp" />
+		<Unit filename="source/level.h" />
+		<Unit filename="source/main.cpp" />
+		<Unit filename="source/mathUtils.cpp" />
+		<Unit filename="source/mathUtils.h" />
+		<Unit filename="source/npc.cpp" />
+		<Unit filename="source/npc.h" />
+		<Unit filename="source/npc000.cpp" />
+		<Unit filename="source/npc000.h" />
+		<Unit filename="source/npc020.cpp" />
+		<Unit filename="source/npc020.h" />
+		<Unit filename="source/npc040.cpp" />
+		<Unit filename="source/npc040.h" />
+		<Unit filename="source/npc060.cpp" />
+		<Unit filename="source/npc060.h" />
+		<Unit filename="source/npc080.cpp" />
+		<Unit filename="source/npc080.h" />
+		<Unit filename="source/npc100.cpp" />
+		<Unit filename="source/npc100.h" />
+		<Unit filename="source/npc120.cpp" />
+		<Unit filename="source/npc120.h" />
+		<Unit filename="source/npc140.cpp" />
+		<Unit filename="source/npc140.h" />
+		<Unit filename="source/npc180.cpp" />
+		<Unit filename="source/npc180.h" />
+		<Unit filename="source/npc200.cpp" />
+		<Unit filename="source/npc200.h" />
+		<Unit filename="source/npc220.cpp" />
+		<Unit filename="source/npc220.h" />
+		<Unit filename="source/npc240.cpp" />
+		<Unit filename="source/npc240.h" />
+		<Unit filename="source/npc280.cpp" />
+		<Unit filename="source/npc280.h" />
+		<Unit filename="source/npc300.cpp" />
+		<Unit filename="source/npc300.h" />
+		<Unit filename="source/npc340.cpp" />
+		<Unit filename="source/npc340.h" />
+		<Unit filename="source/npcAct.cpp" />
+		<Unit filename="source/npcAct.h" />
+		<Unit filename="source/npcCollision.cpp" />
+		<Unit filename="source/npcCollision.h" />
+		<Unit filename="source/org.cpp" />
+		<Unit filename="source/org.h" />
+		<Unit filename="source/player.cpp" />
+		<Unit filename="source/player.h" />
+		<Unit filename="source/playerCollision.cpp" />
+		<Unit filename="source/playerCollision.h" />
+		<Unit filename="source/polarStar.cpp" />
+		<Unit filename="source/polarStar.h" />
+		<Unit filename="source/polarStarShoot.cpp" />
+		<Unit filename="source/render.cpp" />
+		<Unit filename="source/render.h" />
+		<Unit filename="source/resource.h" />
+		<Unit filename="source/script.cpp" />
+		<Unit filename="source/script.h" />
+		<Unit filename="source/shootFunctions.h" />
+		<Unit filename="source/sound.cpp" />
+		<Unit filename="source/sound.h" />
+		<Unit filename="source/spur.cpp" />
+		<Unit filename="source/spur.h" />
+		<Unit filename="source/spurShoot.cpp" />
+		<Unit filename="source/valueview.cpp" />
+		<Unit filename="source/valueview.h" />
+		<Unit filename="source/weapons.cpp" />
+		<Unit filename="source/weapons.h" />
 		<Extensions>
 			<code_completion />
 			<envvars />
