+++ source/npc060.h
@@ -6,18 +6,18 @@
 void npcAct060(npc *NPC);
 void npcAct061(npc *NPC);
 void npcAct062(npc *NPC);
-
+void npcAct063(npc *NPC);
 void npcAct064(npc *NPC);
 void npcAct065(npc *NPC);
 void npcAct066(npc *NPC);
 
 void npcAct070(npc *NPC);
 void npcAct071(npc *NPC);
-
+void npcAct072(npc *NPC);
 void npcAct073(npc *NPC);
 void npcAct074(npc *NPC);
 void npcAct075(npc *NPC);
 void npcAct076(npc *NPC);
-
+void npcAct077(npc *NPC);
 void npcAct078(npc *NPC);
 void npcAct079(npc *NPC);
