+++ source/npcAct.cpp
@@ -84,7 +84,7 @@ npcAct npcActs[361] =
 	&npcAct060,
 	&npcAct061,
 	&npcAct062,
-	&npcActNone,
+	&npcAct063,
 	&npcAct064,
 	&npcAct065,
 	&npcAct066,
@@ -93,7 +93,7 @@ npcAct npcActs[361] =
 	&npcActNone,
 	&npcAct070,
 	&npcAct071,
-	&npcActNone,
+	&npcAct072,
 	&npcAct073,
 	&npcAct074,
 	&npcAct075,
